battery level float now

+++ app/src/main/java/com/example/eddyapp/data/model/GeneralStatusResponse.kt
@@ -10,7 +10,7 @@ data class GeneralStatusResponse(
 data class SystemStatus(
     val connection_mode: String,
     val status: ConnectionStatus,
-    val battery_level: Int,
+    val battery_level: Float,
     val battery_charging: Boolean,
 )
